Update the token and refresh endpoints for open source users of GTM OAuth 2
Add some comments explaining why the code doesn't need to set and check sign-in state parameter, nor to validate the returned ID token.
While iSL (iOS SSO) compiles GTM OAuth 2 in, it doesn't rely on these endpoints.
Web apps are encouraged to fetch a discovery doc to get the endpoints, but for performance and stability we've always hardcoded the endpoints into the native clients.
Build fixes for Xcode 7.