Mac window controller: defer the stopLoading:
In the GTM OAuth2 Mac window controller, defer the stopLoading: message so it's
not sent during the WebView webView:resource:willSendRequest: delegate method
invocation.

This fixes a crash that was reproducible when signing in with DailyMotion. It
does not occur on iOS (as iOS isn't invoking stopLoading on the webview.)

Similar crashes have been reported in GTMOAuth1 on iOS and OS X:
  https://github.com/google/gtm-oauth2/issues/66
  https://groups.google.com/forum/#!searchin/gtm-oauth/stopLoading/gtm-oauth/N6jlOpL9k5g/BXFnNEIudb0J

Also make a few unintentionally global variables static.